Actually control instantiations
