+++ src/ArtificialConduction/ArtificialConductionInst.cc
@@ -1,0 +1,20 @@
+//------------------------------------------------------------------------------
+// Explict instantiation.
+//------------------------------------------------------------------------------
+#include "Spheral/config.hh"
+#include "ArtificialConduction/ArtificialConduction.hh"
+#include "Geometry/Dimension.hh"
+
+namespace Spheral {
+#if defined(SPHERAL_ENABLE_1D)
+  template class ArtificialConduction< Dim< 1 > >;
+#endif
+
+#if defined(SPHERAL_ENABLE_2D)
+  template class ArtificialConduction< Dim< 2 > >;
+#endif
+
+#if defined(SPHERAL_ENABLE_3D)
+  template class ArtificialConduction< Dim< 3 > >;
+#endif
+}
