+++ src/ArtificialConduction/ArtificialConductionPolicyInst.cc
@@ -1,0 +1,17 @@
+#include "Spheral/config.hh"
+#include "ArtificialConduction/ArtificialConductionPolicy.hh"
+#include "Geometry/Dimension.hh"
+
+namespace Spheral {
+#if defined(SPHERAL_ENABLE_1D)
+    template class ArtificialConductionPolicy< Dim< 1 > >;
+#endif
+
+#if defined(SPHERAL_ENABLE_2D)
+    template class ArtificialConductionPolicy< Dim< 2 > >;
+#endif
+
+#if defined(SPHERAL_ENABLE_3D)
+    template class ArtificialConductionPolicy< Dim< 3 > >;
+#endif
+}
