+++ src/ArtificialConduction/CMakeLists.txt
@@ -1,14 +1,22 @@
 include_directories(.)
 
-set(ArtificialConduction_sources
-    ArtificialConduction.cc
-    ArtificialConductionPolicy.cc
-    )
-
 set(ArtificialConduction_headers
     ArtificialConduction.hh
+    ArtificialConductionDecl.hh
+    ArtificialConductionImpl.hh
     ArtificialConductionPolicy.hh
+    ArtificialConductionPolicyDecl.hh
+    ArtificialConductionPolicyImpl.hh
     ) 
+
+set(ArtificialConduction_sources )
+
+if (ENABLE_INSTANTIATIONS)
+  list(APPEND ArtificialConduction_sources
+       ArtificialConductionInst.cc
+       ArtificialConductionPolicyInst.cc
+       )
+endif ()
 
 spheral_add_obj_library(ArtificialConduction SPHERAL_OBJ_LIBS)
